Put notebook stuff into VS Code server
