+++ scripts/notebook/notebook.jl
@@ -1,8 +1,11 @@
 Base.push!(LOAD_PATH, joinpath(@__DIR__, "..", "packages"))
-using VSCodeNotebookServer
+using VSCodeServer
 pop!(LOAD_PATH)
 
 let
-    conn_pipeline = Base.ARGS[1]
-    VSCodeNotebookServer.serve(conn_pipeline)
+    args = [popfirst!(Base.ARGS) for _ in 1:2]
+
+    conn_pipeline, telemetry_pipeline = args[1:2]
+
+    VSCodeServer.serve_notebook(conn_pipeline, crashreporting_pipename=telemetry_pipeline)
 end
