--- scripts/packages/VSCodeNotebookServer/Project.toml
@@ -1,19 +1,0 @@
-name = "VSCodeNotebookServer"
-uuid = "1d4eba16-578d-4c9f-a95f-16f52a066826"
-
-[deps]
-Base64 = "2a0f44e3-6c83-55bd-87e4-b1978d98bd5f"
-Sockets = "6462fe0b-24de-5631-8697-dd941f90decc"
-UUIDs = "cf7118a7-6976-5b1a-9a39-7adc72f591a4"
-Unicode = "4ec0a83e-493e-50e2-b9ac-8f72acf5a8f5"
-Mmap = "a63ad114-7e13-5084-954f-fe012c677804"
-Dates = "ade2ca70-3891-5945-98fb-dc099432e06a"
-
-[extras]
-Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
-
-[compat]
-julia = "1"
-
-[targets]
-test = ["Test"]
